EmployeeRest rewrote with Exception Handler

+++ personnel-management/service/src/main/java/it/academy/service/EmployeeService.java
@@ -1,6 +1,10 @@
 package it.academy.service;
 
+import it.academy.dao.DepartmentDao;
 import it.academy.dao.EmployeeDao;
+import it.academy.exception.*;
+import it.academy.exception.IllegalArgumentException;
+import it.academy.model.Department;
 import it.academy.model.Employee;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,9 +18,19 @@ public class EmployeeService {
     @Autowired
     EmployeeDao employeeDaoImpl;
 
+    @Autowired
+    DepartmentDao departmentDao;
+
+
     @Transactional
-    public Employee getOneEmployee(String id) {
+    public Employee getOneEmployee(String id) throws NotFoundException, IllegalArgumentException {
         Employee employee = employeeDaoImpl.getOneEmployee(id);
+        if(id==null||id==""){
+            throw new IllegalArgumentException("Illegal argument");
+        }
+        if (employee==null){
+            throw new NotFoundException("Employee with that id doesn't exist");
+        }
         return employee;
     }
 
@@ -28,23 +42,45 @@ public class EmployeeService {
 
 
     @Transactional
-    public void deleteEmployee (String id){
+    public void deleteEmployee (String id) throws NotFoundException, IllegalArgumentException {
+        if(id==null||id==""){
+            throw new IllegalArgumentException("Illegal argument");
+        }
+        Employee oneEmployee = getOneEmployee(id);
+        if (oneEmployee==null){
+            throw new NotFoundException("Employee with that id doesn't exist");
+        }
         employeeDaoImpl.delete(id);
     }
 
     @Transactional
-    public void addEmployeeToDepartment(String employeeId, String departmentId){
-        employeeDaoImpl.addEmployeeToDepartment(employeeId,departmentId);
+    public void addEmployeeToDepartment(String employeeId, String departmentId)
+            throws NotFoundException, IllegalArgumentException {
+//        employeeDaoImpl.addEmployeeToDepartment(employeeId,departmentId);
+        Employee oneEmployee = getOneEmployee(employeeId);
+        Department oneDepartment = departmentDao.getOneDepartment(departmentId);
+        employeeDaoImpl.addEmployeeToDepartment(oneEmployee, oneDepartment);
+
     }
 
     @Transactional
-    public void removeEmployeeFromDepartment(String id){
+    public void removeEmployeeFromDepartment(String id) throws NotFoundException, IllegalArgumentException {
+        if(id==null||id==""){
+            throw new IllegalArgumentException("Illegal argument");
+        }
+        Employee oneEmployee = getOneEmployee(id);
+        if (oneEmployee==null){
+            throw new NotFoundException("Employee with that id doesn't exist");
+        }
         employeeDaoImpl.removeEmployeeFromDepartment(id);
     }
 
     @Transactional
-    public List<Employee> getAllEmployeeWithoutDepartment(){
+    public List<Employee> getAllEmployeeWithoutDepartment() throws NotFoundException {
         List<Employee> allEmployeeWithoutDepartment = employeeDaoImpl.getAllEmployeeWithoutDepartment();
+        if (allEmployeeWithoutDepartment==null){
+            throw new NotFoundException("Employee without department doesn't exist");
+        }
         return allEmployeeWithoutDepartment;
     }
